% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Country Population}
\usage{
CountryPopulation(CountryInput)
}
\arguments{
\item{CountryInput}{A string that spells out a country}
}
\value{
A graph that represents the country's population from 1950-2020
}
\description{
Takes in a character string that is a country name and returns a graph of
the population of that country over time.
}
