% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldCupDocumentation.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{A data frame containing World Cup data including the year, hosting country,
total attendance, and average attendance}
\format{
a data frame with 22 observations
\describe{
  \item{WorldCup}{The name and year of a World Cup and its hosting country.}
  \item{Matches}{Matches played at that World Cup.}
  \item{Totalattendance}{The total attendance for that game.}
  \item{Averageattendance}{The average attendance for that game.}
}
}
\source{
\url{https://en.wikipedia.org/wiki/FIFA_World_Cup}
}
\usage{
World_Cup
}
\description{
A data frame containing World Cup data including the year, hosting country,
total attendance, and average attendance
}
\keyword{datasets}
