% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopDocumentation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{A data frame containing UN data relating to the world population of countries
from 1950-2020}
\format{
a data frame with 235 observations
\describe{
  \item{CountryName}{The name of all countries throughout 1950 to 2020.}
  \item{Years}{The total country populations within those years.}
}
}
\source{
\xlsx{World_Population.xlsx}
}
\usage{
WorldPopulation
}
\description{
A data frame containing UN data relating to the world population of countries
from 1950-2020
}
\keyword{datasets}
